configuration is it's own member of the workspace
This separates concerns, and reduces the chances of
unintended circular dependencies.

`cli` and `lib` both depend on `config`
