+++ cli/src/lib.rs
@@ -4,6 +4,7 @@ use std::sync::Arc;
 
 use log::{error, info};
 
+use zingoconfig::Network;
 use zingolib::lightclient::lightclient_config::LightClientConfig;
 use zingolib::{commands, lightclient::LightClient};
 
@@ -232,7 +233,7 @@ pub fn attempt_recover_seed(_password: Option<String>) {
     // Create a Light Client Config in an attempt to recover the file.
     let _config = LightClientConfig {
         server: "0.0.0.0:0".parse().unwrap(),
-        chain: zingolib::lightclient::lightclient_config::Network::Mainnet,
+        chain: zingoconfig::Network::Mainnet,
         monitor_mempool: false,
         anchor_offset: [0u32; 5],
         data_dir: None,
