+++ lib/Cargo.toml
@@ -57,6 +57,7 @@ zcash_client_backend = { git = "https://github.com/zcash/librustzcash", rev = "7
 zcash_proofs = { git = "https://github.com/zcash/librustzcash", rev = "73d9395c9d3c5fa81fc7becd363a2c1a51772a76", features = ["multicore"]}
 zcash_encoding = { git = "https://github.com/zcash/librustzcash", rev = "73d9395c9d3c5fa81fc7becd363a2c1a51772a76"}
 zcash_note_encryption = { git = "https://github.com/zcash/librustzcash", rev = "73d9395c9d3c5fa81fc7becd363a2c1a51772a76", features = ["pre-zip-212"]}
+zingoconfig = { path = "../config" }
 
 [dev-dependencies]
 portpicker = "0.1.0"
