+++ lib/src/wallet/keys/transparent.rs
@@ -344,8 +344,7 @@ mod test {
 
     #[test]
     fn tkey_encode_decode() {
-        let config =
-            LightClientConfig::create_unconnected(crate::lightclient::lightclient_config::Network::FakeMainnet, None);
+        let config = LightClientConfig::create_unconnected(zingoconfig::Network::FakeMainnet, None);
 
         for _i in 0..10 {
             let mut b = [0u8; 32];
